Let's try putting this new code in the "right" place ...
